Re-write to exporter
This commit removes any integration with the billing database. The
metrics will be provided via an exporter. This integrates more nicely
with the rest of the billing system.

Metrics that are provided once a day, like for object storage, are
only scraped after a specified time. Then the value will be divided by
the rest of hours for the given day.

Exoscale DBaaS metrics are provided as is.

The metrics exported have the name `appcat:raw:billing` as they can't
be used directly for the billing. A set of prometheus rules will do
some aggregation and properly export them as `appcat:billing`. These
rules are part of the component and not included in this commit.